+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## [2.0.0] - 2023-10-18
 ### Changed
 - Removed dependencies on `http`, `bluebird`, `request`
+- Use customgrouping3 instead of customgrouping for categorization
+
+## Removed
+- Possibility to use a query string
 
 ## [1.4.1] - 2022-12-12
 ### Changed